+++ tests/test_doctestplus.py
@@ -596,3 +596,28 @@ if LooseVersion('4.3.0') <= LooseVersion(pytest.__version__):
         testdir.inline_run(
             '--doctest-plus', '--doctest-rst', '--ignore-glob', '*.rst'
         ).assertoutcome(passed=2)
+
+
+def test_doctest_only(testdir, makepyfile, maketestfile, makerstfile):
+    # regular python files with doctests
+    makepyfile(p1='>>> 1 + 1\n2')
+    makepyfile(p2='>>> 1 + 1\n3')
+    # regular test files
+    maketestfile(test_1='foo')
+    maketestfile(test_2='bar')
+    # rst files
+    makerstfile(r1='>>> 1 + 1\n2')
+    makerstfile(r3='>>> 1 + 1\n3')
+    makerstfile(r2='>>> 1 + 2\n3')
+
+    # regular tests
+    testdir.inline_run().assertoutcome(passed=2)
+    # regular + doctests
+    testdir.inline_run("--doctest-plus").assertoutcome(passed=3, failed=1)
+    # regular + doctests + doctest in rst files
+    testdir.inline_run("--doctest-plus", "--doctest-rst").assertoutcome(passed=5, failed=2)
+    # only doctests in python files, implicit usage of doctest-plus
+    testdir.inline_run("--doctest-only").assertoutcome(passed=1, failed=1)
+    # only doctests in python files
+    testdir.inline_run("--doctest-only", "--doctest-rst").assertoutcome(passed=3, failed=2)
+